requirements.txt
numpy
scipy
sympy
matplotlib
pandas
pyyaml
pillow
